+++ includes/class-tmsm-admin-cleanup.php
@@ -163,8 +163,11 @@ class Tmsm_Admin_Cleanup {
 		// Plugin Updates: Disable for some plugins
 		$this->loader->add_filter( 'site_transient_update_plugins', $plugin_admin, 'site_transient_update_plugins_disable_specific', 50, 3 );
 
-		// Core Updates: disable wp_version_check single event creation
-		$this->loader->add_filter( 'pre_schedule_event', $plugin_admin, 'pre_schedule_event_disableversioncheck', 10, 3 );
+		// Cores Updates: disable auto updates after wp_version_check
+		remove_action('wp_maybe_auto_update', 'wp_maybe_auto_update');
+
+		// Core Updates: disable wp_version_check single event creation (removed because it prevents core security from being applied with a ttl
+		//$this->loader->add_filter( 'pre_schedule_event', $plugin_admin, 'pre_schedule_event_disableversioncheck', 10, 3 );
 
 		// Dashboard
 		$this->loader->add_action( 'admin_init', $plugin_admin, 'remove_dashboard_boxes');