+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+* *Revert* Core Updates: disable `wp_version_check` single event creation (forced ttl to 3600 prevented the weekly schedule of the updates)
+* Instead disable `wp_maybe_auto_update` action to prevent `wp_update_plugins` from being triggered after `wp_version_check`
+
 ### 1.5.0: January 6th, 2022
-* Core Updates: disable wp_version_check single event creation (forced ttl to 3600 prevented the weekly schedule of the updates)
+* Core Updates: disable `wp_version_check` single event creation (forced ttl to 3600 prevented the weekly schedule of the updates)
 
 ### 1.4.9: December 10th, 2021
 * WooCommerce: remove flickering screen because of spinner